requirements.txt
PySide6
pdfminer.six
pyhwp
olefile
pyinstaller
